base/docker/requirements.txt
torch>=2.0
open_clip_torch>=2.23
git+https://github.com/openai/CLIP.git
dbeval 
slack_sdk
pillow